fix stop error test fix
